Fix authenticate router
