+++ src/App.jsx
@@ -1,5 +1,4 @@
 import Header from "components/header";
-import SignForm from "pages/sign";
 import { useSelector } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import Routes from "routes/Routes";
@@ -12,28 +11,17 @@ function App() {
   // detect when user close page to change active => offline
   window.addEventListener("beforeunload", (e) => {
     e.preventDefault();
-    
-    if(user){
-      updateInActiveUser(user.uid)
-    };
+
+    if (user) {
+      updateInActiveUser(user.uid);
+    }
   });
 
   return (
-    <>
-      {!user ? (
-        <SignForm />
-      ) : (
-        <BrowserRouter>
-          {/* Header */}
-          <Header />
-          {/* Body */}
-          <Routes />
-          {/* sidebar */}
-          {/* feeds */}
-          {/* widgets */}
-        </BrowserRouter>
-      )}
-    </>
+    <BrowserRouter>
+      <Header />
+      <Routes />
+    </BrowserRouter>
   );
 }
